Move query to another file
